% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTask.R
\name{addTask}
\alias{addTask}
\title{Add a new task to a \code{taskList}}
\usage{
addTask(taskList, taskId, taskCreator = NULL, taskDescription = NULL,
  taskDateCreated = NULL, taskDeadline = NULL)
}
\arguments{
\item{taskId}{the taskId}

\item{taskCreator}{the name of the creator of the task}

\item{taskDescription}{a description of the task to do}

\item{taskDateCreated}{the date on which the task is created}

\item{taskDeadline}{the date on which the task has to be finished}
}
\description{
Add a new task to a \code{taskList}
}
