% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjects.R
\name{listProjects}
\alias{listProjects}
\title{List all file names in a specified \code{path} which match a certain \code{fileNamePattern}}
\usage{
listProjects(fileNamePattern, path = getwd(), full.names = FALSE)
}
\description{
List all file names in a specified \code{path} which match a certain \code{fileNamePattern}
}
